samantha_ai_assistant/apps/mcp-server/generate_with_automcp.py
"""
AutoMCP Integration Script for Samantha AI
Automatically generates MCP server code from existing APIs
"""

import yaml
import json
import logging
from pathlib import Path
from typing import Dict, Any, Optional

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class SamanthaAIAutoMCP:
    """AutoMCP integration for Samantha AI"""

    def __init__(self, config_path: str = "automcp.yaml"):
        self.config_path = config_path
        self.config = self.load_config()
        logger.info("Samantha AI AutoMCP initialized")

    def load_config(self) -> Dict[str, Any]:
        """Load AutoMCP configuration"""
        try:
            with open(self.config_path, 'r') as f:
                config = yaml.safe_load(f)
            logger.info(f"Loaded AutoMCP config from {self.config_path}")
            return config
        except Exception as e:
            logger.error(f"Error loading config: {str(e)}")
            return {}

    def generate_mcp_server_code(self) -> str:
        """Generate MCP server code from configuration"""
        try:
            logger.info("Generating MCP server code...")

            # Generate imports
            imports = self._generate_imports()

            # Generate server class
            server_class = self._generate_server_class()

            # Generate tools
            tools = self._generate_tools()

            # Generate resources
            resources = self._generate_resources()

            # Generate main function
            main_function = self._generate_main_function()

            # Combine all code
            code = f'''"""
Samantha AI MCP Server - Auto-Generated
Voice-controlled AI assistant with system automation
"""

{imports}

{server_class}

{tools}

{resources}

{main_function}
'''

            logger.info("MCP server code generated successfully")
            return code

        except Exception as e:
            logger.error(f"Error generating MCP server code: {str(e)}")
            return ""

    def _generate_imports(self) -> str:
        """Generate import statements"""
        return '''import asyncio
import logging
from typing import Dict, Any, Optional
from datetime import datetime
import base64

from mcp import Server, StdioServerParameters
from mcp.server import Session
from mcp.server.models import InitializationOptions
from mcp.server.stdio import StdioServerTransport

# Import existing services
from backend.voice_processor import VoiceProcessor, VoiceCommand
from backend.system_automation import SystemAutomation

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)'''

    def _generate_server_class(self) -> str:
        """Generate main server class"""
        return '''class SamanthaAIMCPServer(Server):
    """Main MCP server for Samantha AI voice assistant - Auto-Generated"""

    def __init__(self):
        super().__init__(
            name="samantha-ai-voice-assistant",
            version="1.0.0"
        )

        # Initialize existing services
        self.voice_processor = VoiceProcessor()
        self.system_automation = SystemAutomation()

        logger.info("Samantha AI MCP Server initialized")

    async def initialize(self, session: Session, params: InitializationOptions):
        """Initialize MCP server with all tools and resources"""
        logger.info("Initializing MCP server...")

        # Register tools and resources will be added here
        logger.info("MCP server initialization complete")'''

    def _generate_tools(self) -> str:
        """Generate tool methods from configuration"""
        tools_code = []

        for tool in self.config.get('tools', []):
            tool_name = tool.get('name', '')
            description = tool.get('description', '')
            input_schema = tool.get('input_schema', {})
            output_schema = tool.get('output_schema', {})

            tool_code = f'''
    async def {tool_name}(self, **kwargs) -> Dict[str, Any]:
        """{description}"""
        try:
            logger.info(f"Executing {tool_name} with params: {{kwargs}}")

            # TODO: Implement actual tool logic
            # This is a placeholder for auto-generated tool

            return {{
                "success": True,
                "message": "Tool {tool_name} executed successfully",
                "data": kwargs,
                "timestamp": datetime.now().isoformat()
            }}
        except Exception as e:
            logger.error(f"Error in {tool_name}: {{str(e)}}")
            return {{
                "success": False,
                "error": str(e)
            }}'''

            tools_code.append(tool_code)

        return '\n'.join(tools_code)

    def _generate_resources(self) -> str:
        """Generate resource methods from configuration"""
        resources_code = []

        for resource in self.config.get('resources', []):
            resource_name = resource.get('name', '')
            description = resource.get('description', '')
            schema = resource.get('schema', {})

            resource_code = f'''
    async def get_{resource_name}(self) -> Dict[str, Any]:
        """{description}"""
        try:
            logger.info(f"Getting {resource_name}")

            # TODO: Implement actual resource logic
            # This is a placeholder for auto-generated resource

            return {{
                "success": True,
                "data": {{
                    "resource": "{resource_name}",
                    "description": "{description}",
                    "timestamp": datetime.now().isoformat()
                }}
            }}
        except Exception as e:
            logger.error(f"Error getting {resource_name}: {{str(e)}}")
            return {{
                "success": False,
                "error": str(e)
            }}'''

            resources_code.append(resource_code)

        return '\n'.join(resources_code)

    def _generate_main_function(self) -> str:
        """Generate main function"""
        return '''async def main():
    """Main entry point for MCP server"""
    logger.info("Starting Samantha AI MCP Server...")

    server = SamanthaAIMCPServer()

    # Run with stdio transport
    async with StdioServerTransport() as (read_stream, write_stream):
        await server.run(
            StdioServerParameters(
                read_stream=read_stream,
                write_stream=write_stream
            )
        )

if __name__ == "__main__":
    asyncio.run(main())'''

    def generate_api_mapping(self) -> Dict[str, Any]:
        """Generate API endpoint to MCP tool mapping"""
        mapping = {}

        for endpoint in self.config.get('api_endpoints', []):
            path = endpoint.get('path', '')
            method = endpoint.get('method', 'GET')
            tool_name = endpoint.get('tool_name', '')
            resource_name = endpoint.get('resource_name', '')

            if tool_name:
                mapping[path] = {
                    'method': method,
                    'type': 'tool',
                    'name': tool_name
                }
            elif resource_name:
                mapping[path] = {
                    'method': method,
                    'type': 'resource',
                    'name': resource_name
                }

        return mapping

    def save_generated_code(self, output_path: str = "generated_mcp_server.py"):
        """Save generated MCP server code to file"""
        try:
            code = self.generate_mcp_server_code()

            with open(output_path, 'w') as f:
                f.write(code)

            logger.info(f"Generated MCP server code saved to {output_path}")

            # Also save API mapping
            mapping = self.generate_api_mapping()
            mapping_path = output_path.replace('.py', '_mapping.json')

            with open(mapping_path, 'w') as f:
                json.dump(mapping, f, indent=2)

            logger.info(f"API mapping saved to {mapping_path}")

            return True

        except Exception as e:
            logger.error(f"Error saving generated code: {str(e)}")
            return False

    def validate_configuration(self) -> bool:
        """Validate AutoMCP configuration"""
        try:
            required_keys = ['server', 'tools', 'resources']

            for key in required_keys:
                if key not in self.config:
                    logger.error(f"Missing required key: {key}")
                    return False

            # Validate server configuration
            server_config = self.config.get('server', {})
            required_server_keys = ['name', 'version', 'description']

            for key in required_server_keys:
                if key not in server_config:
                    logger.error(f"Missing required server key: {key}")
                    return False

            # Validate tools
            for tool in self.config.get('tools', []):
                required_tool_keys = ['name', 'description', 'input_schema', 'output_schema']

                for key in required_tool_keys:
                    if key not in tool:
                        logger.error(f"Missing required tool key: {key}")
                        return False

            # Validate resources
            for resource in self.config.get('resources', []):
                required_resource_keys = ['name', 'description', 'schema']

                for key in required_resource_keys:
                    if key not in resource:
                        logger.error(f"Missing required resource key: {key}")
                        return False

            logger.info("Configuration validation passed")
            return True

        except Exception as e:
            logger.error(f"Configuration validation error: {str(e)}")
            return False

def main():
    """Main function to generate MCP server"""
    logger.info("Starting AutoMCP generation...")

    # Initialize AutoMCP
    automcp = SamanthaAIAutoMCP()

    # Validate configuration
    if not automcp.validate_configuration():
        logger.error("Configuration validation failed")
        return False

    # Generate and save code
    if automcp.save_generated_code():
        logger.info("✅ MCP server generation completed successfully!")
        return True
    else:
        logger.error("❌ MCP server generation failed")
        return False

if __name__ == "__main__":
    success = main()
    exit(0 if success else 1)
