samantha_ai_assistant/apps/mcp-server/working_mcp_server.py
"""
Working Samantha AI MCP Server
Voice-controlled AI assistant with system automation
"""

import asyncio
import base64
import functools
import io
import logging
import os
from typing import Dict, Any, Optional
from datetime import datetime

from dotenv import load_dotenv
import openai
import elevenlabs

from mcp import StdioServerParameters
from mcp.server import Server
from mcp.types import CallToolResult, TextContent, AudioContent

# Load environment variables from .env file
load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class SamanthaAIMCPServer(Server):
    """Main MCP server for Samantha AI voice assistant"""

    def __init__(self):
        super().__init__(
            name="samantha-ai-voice-assistant",
            version="1.0.0"
        )

        # Initialize API clients
        self.openai_client = openai.AsyncClient(api_key=os.getenv("OPENAI_API_KEY"))
        elevenlabs_api_key = os.getenv("ELEVENLABS_API_KEY")
        if elevenlabs_api_key:
            elevenlabs.set_api_key(elevenlabs_api_key)

        logger.info("Samantha AI MCP Server initialized")

        # Register tools
        self._register_tools()

    def _register_tools(self):
        """Register all MCP tools"""

        # Voice processing tools
        @self.call_tool()
        async def speech_to_text(self, audio_data: str, language: str = 'en-US') -> CallToolResult:
            """Convert speech to text using OpenAI Whisper"""
            try:
                logger.info(f"Processing speech to text with language: {language}")

                # The audio_data is base64 encoded. Decode it and wrap in a file-like object.
                try:
                    decoded_audio = base64.b64decode(audio_data)
                    audio_file = io.BytesIO(decoded_audio)
                    # The whisper API requires a file name.
                    audio_file.name = "input.wav"
                except (base64.binascii.Error, TypeError) as e:
                    logger.error(f"Invalid base64 audio data: {e}")
                    return CallToolResult(
                        content=[TextContent(type="text", text=f"Invalid audio format: {e}")],
                        isError=True
                    )

                transcript = await self.openai_client.audio.transcriptions.create(
                    model="whisper-1",
                    file=audio_file,
                    language=language
                )

                return CallToolResult(
                    content=[TextContent(type="text", text=transcript.text)]
                )
            except Exception as e:
                logger.error(f"Speech to text error: {str(e)}")
                return CallToolResult(
                    content=[TextContent(type="text", text=f"Error in speech to text: {e}")],
                    isError=True
                )

        @self.call_tool()
        async def text_to_speech(self, text: str, voice: str = 'alloy') -> CallToolResult:
            """Convert text to speech using ElevenLabs"""
            try:
                logger.info(f"Converting text to speech: {text[:50]}...")
                if not os.getenv("ELEVENLABS_API_KEY"):
                    logger.warning("ELEVENLABS_API_KEY not set, skipping TTS.")
                    return CallToolResult(
                        content=[TextContent(type="text", text="ElevenLabs API key not configured.")],
                        isError=True
                    )

                # elevenlabs.generate is synchronous, so run it in an executor
                loop = asyncio.get_running_loop()
                audio_bytes = await loop.run_in_executor(
                    None, functools.partial(elevenlabs.generate, text=text, voice=voice, model="eleven_multilingual_v2")
                )

                audio_b64 = base64.b64encode(b"".join(audio_bytes)).decode('utf-8')
                return CallToolResult(content=[AudioContent(type="audio", data=audio_b64, mimeType="audio/mpeg")])
            except Exception as e:
                logger.error(f"Text to speech error: {str(e)}")
                return CallToolResult(content=[TextContent(type="text", text=f"Error in TTS: {e}")], isError=True)

        @self.call_tool()
        async def intent_classification(text: str) -> Dict[str, Any]:
            """Classify user intent from text"""
            try:
                logger.info(f"Classifying intent for text: {text[:50]}...")

                # TODO: Implement actual intent classification
                # For now, return mock data

                return {
                    "success": True,
                    "intent": "system_control",
                    "confidence": 0.85,
                    "entities": {"action": "volume_up"},
                    "text": text
                }
            except Exception as e:
                logger.error(f"Intent classification error: {str(e)}")
                return {
                    "success": False,
                    "error": str(e)
                }

        # System automation tools
        @self.call_tool()
        async def file_operation(operation: str, params: Dict[str, Any]) -> Dict[str, Any]:
            """Execute file system operations"""
            try:
                logger.info(f"Executing file operation: {operation}")

                # TODO: Implement actual file operations
                # For now, return mock data

                return {
                    "success": True,
                    "message": f"File operation '{operation}' executed successfully",
                    "data": params,
                    "error": None,
                    "execution_time": 0.1
                }
            except Exception as e:
                logger.error(f"File operation error: {str(e)}")
                return {
                    "success": False,
                    "error": str(e)
                }

        @self.call_tool()
        async def app_control(action: str, app_name: str) -> Dict[str, Any]:
            """Control applications"""
            try:
                logger.info(f"Executing app control: {action} for {app_name}")

                # TODO: Implement actual app control
                # For now, return mock data

                return {
                    "success": True,
                    "message": f"App '{app_name}' {action} successful",
                    "data": {"app": app_name, "action": action},
                    "error": None,
                    "execution_time": 0.2
                }
            except Exception as e:
                logger.error(f"App control error: {str(e)}")
                return {
                    "success": False,
                    "error": str(e)
                }

        @self.call_tool()
        async def system_control(action: str, params: Dict[str, Any] = None) -> Dict[str, Any]:
            """Control system settings"""
            try:
                logger.info(f"Executing system control: {action}")

                # TODO: Implement actual system control
                # For now, return mock data

                return {
                    "success": True,
                    "message": f"System control '{action}' executed successfully",
                    "data": params or {},
                    "error": None,
                    "execution_time": 0.15
                }
            except Exception as e:
                logger.error(f"System control error: {str(e)}")
                return {
                    "success": False,
                    "error": str(e)
                }

        @self.call_tool()
        async def browser_automation(browser: str, action: str, params: Dict[str, Any] = None) -> Dict[str, Any]:
            """Automate browser operations"""
            try:
                logger.info(f"Executing browser automation: {browser} - {action}")

                # TODO: Implement actual browser automation
                # For now, return mock data

                return {
                    "success": True,
                    "message": f"Browser '{browser}' {action} successful",
                    "data": {"browser": browser, "action": action, "params": params},
                    "error": None,
                    "execution_time": 0.3
                }
            except Exception as e:
                logger.error(f"Browser automation error: {str(e)}")
                return {
                    "success": False,
                    "error": str(e)
                }

        @self.call_tool()
        async def execute_automation(command: str, params: Dict[str, Any] = None) -> Dict[str, Any]:
            """Execute any automation command"""
            try:
                logger.info(f"Executing automation command: {command}")

                # TODO: Implement actual automation
                # For now, return mock data

                return {
                    "success": True,
                    "message": f"Automation command '{command}' executed successfully",
                    "data": {"command": command, "params": params},
                    "error": None,
                    "execution_time": 0.25
                }
            except Exception as e:
                logger.error(f"Automation execution error: {str(e)}")
                return {
                    "success": False,
                    "error": str(e)
                }

        # Health check tool
        @self.call_tool()
        async def get_system_health() -> Dict[str, Any]:
            """Get system health status"""
            try:
                # TODO: Implement actual health checks
                # For now, return mock data

                return {
                    "status": "healthy",
                    "voice_processor": {"status": "healthy", "uptime": "2h 15m"},
                    "system_automation": {"status": "healthy", "uptime": "2h 15m"},
                    "timestamp": datetime.now().isoformat()
                }
            except Exception as e:
                logger.error(f"Health check error: {str(e)}")
                return {
                    "status": "error",
                    "error": str(e),
                    "timestamp": datetime.now().isoformat()
                }

async def main():
    """Main entry point for MCP server"""
    logger.info("Starting Samantha AI MCP Server...")

    if not os.getenv("OPENAI_API_KEY"):
        logger.error("FATAL: OPENAI_API_KEY environment variable is not set.")
        return

    server = SamanthaAIMCPServer()

    # The server will listen on stdin/stdout for messages
    logger.info("✅ MCP Server ready for use!")
    await server.listen(StdioServerParameters())

if __name__ == "__main__":
    asyncio.run(main())
