samantha_ai_assistant/apps/mcp-server/requirements.txt
# MCP Server Dependencies
mcp>=1.12.0
automcp>=0.1.0

# Web Framework
fastapi>=0.116.0
uvicorn>=0.35.0
websockets>=15.0.0

# AI and Voice Processing
openai>=1.97.0
elevenlabs>=2.8.0

# HTTP and Async
aiohttp>=3.12.0
python-multipart>=0.0.20

# Configuration and Utilities
pyyaml>=6.0.0
python-dotenv>=1.1.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
